add workaround for Bun.build HTML caching issue
Bun.build does not seem to change the etag header of HTML files when
changing the hash part of asset filenames inside that HTML file because
the content of the original HTML file (before injecting the asset URLs)
didn't change. Because of this the browser loads the cached HTML file
and tries to load the old assets, which lead to 404s.

as a workaround we just inject the current unix timestamp at buildtime
converted to base 36 (to make it a short string like "mfqidl1z") as the
etag for HTML files (we know they won't change after building).

As previously reported
[here](discord://-/channels/912031225894539264/1217525540660183241/1413927866982207678)

Change-Id: I23907dd4fb82c5333f337c04c5eb11d323cda8c5