divide length by 1.5 when DT is detected
(plus a bunch of a refactoring but eh who cares about that)

one issue with this is that DT isn't detected until the round has actually started and the players are in-game, so the length doesn't update until then. Perhaps falling back to the mappool is an idea for that.

Change-Id: I1794e0618216f5b5b27daecec9b50c519485e31f